Adding in WebFOCUS Functions
Snippets for the WebFOCUS functions as of release 8202M. I used the same prefix ('wff'=WebFOCUS Function) so all that you need to do it type wff in VS Code and a list of all the functions will appear. 

I also added in "scope": "focexec" to the existing snippets so that they are only available in .fex files.